Added basic material-like design to app.

+++ expenses-app/src/components/TransactionListRow/index.jsx
@@ -1,42 +1,43 @@
-import React, { useContext } from 'react';
+import React, { useContext } from "react";
 
-import { TransactionContext } from '../../contexts/TransactionContext';
-import { CurrencyContext } from '../../contexts/CurrencyContext';
+import { TransactionContext } from "../../contexts/TransactionContext";
+import { CurrencyContext } from "../../contexts/CurrencyContext";
+
+import "./styles.css";
 
 const TransactionListRow = ({ transaction, handleEdit }) => {
   const { deleteTransaction } = useContext(TransactionContext);
   const { getCurrencyName } = useContext(CurrencyContext);
 
   const handleDelete = () => {
-    var messageString = `are you sure you want to delete transaction ${transaction.title}?`
-    const confirmDelete = window.confirm(messageString)
-    if (confirmDelete)
-      deleteTransaction(transaction.id);
+    var messageString = `are you sure you want to delete transaction ${transaction.title}?`;
+    const confirmDelete = window.confirm(messageString);
+    if (confirmDelete) deleteTransaction(transaction.id);
   };
 
   // check if user is on HomePage
-  const isHomePAge = window.location.pathname === '/';
+  const isHomePage = window.location.pathname === "/";
 
   return (
-    <tr>
-      <td>{transaction.tmpId}</td>
-      <td>{transaction.title}</td>
-      <td>{transaction.categoryName}</td>
-      <td>{transaction.date}</td>
-      <td>{transaction.amount}</td>
+    <tr className="transaction-row">
+      <td className="transaction-data">{transaction.tmpId}</td>
+      <td className="transaction-data">{transaction.title}</td>
+      <td className="transaction-data">{transaction.categoryName}</td>
+      <td className="transaction-data">{transaction.date}</td>
+      <td className="transaction-data">{transaction.amount}</td>
       <td
+        className="transaction-data"
         title={getCurrencyName(transaction.currencyCode)}
       >
         {transaction.currencyCode}
       </td>
-      <td>{transaction.description}</td>
-      {
-      isHomePAge ? null : 
-        <td>
+      <td className="transaction-data">{transaction.description}</td>
+      {isHomePage ? null : (
+        <td className="transaction-actions">
           <button onClick={() => handleEdit(transaction.id)}>Edit</button>
           <button onClick={handleDelete}>Delete</button>
         </td>
-      }
+      )}
     </tr>
   );
 };
